+++ src/main/java/dev/walgo/walib/PackageMetaInfo.java
@@ -5,7 +5,8 @@ import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
-import java.time.LocalDate;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.jar.Attributes;
 import java.util.jar.JarInputStream;
@@ -16,15 +17,14 @@ import org.slf4j.LoggerFactory;
 /**
  * Read meta information (MANIFEST.MF) from package, which contains given class.
  *
- * @author Walery Wysotsky <dev@wysotsky.info>
+ * @author Walery Wysotsky {@literal <dev@wysotsky.info>}
  */
 public class PackageMetaInfo {
 
-  /** show stacktrace, when exception catched. */
+  /** show stacktrace, when exception caught. */
   public static boolean SHOW_STACKTRACE = true;
 
   private static final Logger LOG = LoggerFactory.getLogger(PackageMetaInfo.class);
-  private static final String DEFAULT_DATE_FORMAT = "dd-MM-yyyy HH:mm:ss";
   private static final String MANIFEST_PATH = "/META-INF/MANIFEST.MF";
   private static final String CLASS_EXTENSION = ".class";
 
@@ -36,10 +36,18 @@ public class PackageMetaInfo {
 
   private final String dateFormat;
 
+  /**
+   * Default constructor.
+   */
   public PackageMetaInfo() {
-    this.dateFormat = DEFAULT_DATE_FORMAT;
+    this.dateFormat = null;
   }
 
+  /**
+   * Constructor with defined date format
+   * 
+   * @param dateFormat build dat format (ISO by default)
+   */
   public PackageMetaInfo(String dateFormat) {
     this.dateFormat = dateFormat;
   }
@@ -47,7 +55,7 @@ public class PackageMetaInfo {
   /**
    * Read meta info from package, which contains given class.
    *
-   * @param clazz
+   * @param clazz Base class for extract information
    */
   public void read(Class<?> clazz) {
     String className = clazz.getSimpleName() + CLASS_EXTENSION;
@@ -80,7 +88,7 @@ public class PackageMetaInfo {
   /**
    * Read meta info from package.
    *
-   * @param jarFile
+   * @param jarFile JAR file for read info
    */
   public void read(File jarFile) {
     try {
@@ -106,25 +114,45 @@ public class PackageMetaInfo {
     apiVersion = attributes.getValue("Specification-Version");
   }
 
-  /** @return the title */
+  /**
+   * Gets title from manifest.
+   * 
+   * @return the title
+   */
   public String getTitle() {
     return title;
   }
 
-  /** @return the version */
+  /**
+   * Gets version from manifest.
+   * 
+   * @return the version
+   */
   public String getVersion() {
     return version;
   }
 
-  /** @return the builtDate */
+  /**
+   * Gets build date and time as is (as string).
+   * 
+   * @return the builtDate
+   */
   public String getBuiltDateStr() {
     return builtDate;
   }
 
-  /** @return the builtDate */
-  public LocalDate getBuiltDate() {
+  /**
+   * Gets build date and time.
+   * 
+   * @return the builtDate
+   */
+  public ZonedDateTime getBuiltDate() {
     try {
-      return LocalDate.parse(builtDate, DateTimeFormatter.ofPattern(dateFormat));
+      if (dateFormat != null) {
+        return ZonedDateTime.parse(builtDate, DateTimeFormatter.ofPattern(dateFormat));
+      } else {
+        return ZonedDateTime.parse(builtDate);
+      }
     } catch (Throwable ex) {
       if (SHOW_STACKTRACE) {
         LOG.error(
@@ -133,34 +161,57 @@ public class PackageMetaInfo {
       } else {
         LOG.error("Can't parse built date [{}] with format [{}]", builtDate, dateFormat);
       }
-      return LocalDate.now();
+      return ZonedDateTime.now(ZoneId.systemDefault());
     }
   }
 
-  /** @return the author */
+  /**
+   * Gets package author.
+   * 
+   * @return the author
+   */
   public String getAuthor() {
     return author;
   }
 
-  /** @return the apiVersion */
+  /**
+   * Gets API version (if exists).
+   * 
+   * @return the apiVersion
+   */
   public String getAPIVersion() {
     return apiVersion;
   }
 
+  /**
+   * Package version full info.
+   * 
+   * @return package version generated string
+   */
   public String versionFullString() {
-    String str =
-        String.format(
-            "%s, version %s, built %s by %s",
-            getTitle(), getVersion(), getBuiltDate(), getAuthor());
+    String str = String.format(
+        "%s, version %s, built %s by %s",
+        getTitle(), getVersion(), getBuiltDate(), getAuthor());
     return str;
   }
 
+  /**
+   * Build package meta data based on specific class.
+   * 
+   * @param clazz Class for choose package
+   * @return package meta info
+   */
   public static PackageMetaInfo build(Class<?> clazz) {
     PackageMetaInfo meta = new PackageMetaInfo();
     meta.read(clazz);
     return meta;
   }
 
+  /**
+   * Build package meta data based on current class.
+   * 
+   * @return package meta info
+   */
   public static PackageMetaInfo build() {
     PackageMetaInfo meta = new PackageMetaInfo();
     meta.read(PackageMetaInfo.class);