+++ src/test/java/dev/walgo/walib/CommonsTest.java
@@ -5,13 +5,11 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 
 import org.junit.jupiter.api.Test;
 
-/** @author Walery Wysotsky <dev@wysotsky.info> */
 public class CommonsTest {
 
   private static final String TEST_FIRST = "first";
   private static final String TEST_SECOND = "second";
 
-  /** Test of getTitle method, of class MetaInfo. */
   @Test
   public void testCoalesceFirst() {
     String result = Commons.coalesce(TEST_FIRST, null, TEST_SECOND);