+++ src/test/java/dev/walgo/walib/ResourceLoaderTest.java
@@ -9,14 +9,12 @@ import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
 
-/** @author Walery Wysotsky <dev@wysotsky.info> */
 public class ResourceLoaderTest {
 
   static final String TEST_JAR = "build/resources/test/test.jar";
   static final String TEST_ZIP = "build/resources/test/test.zip";
   static final String TEST_DIR = "build/resources/test/test.dir";
 
-  /** Test of load method, of class ClasspathResourceUtil. */
   @Test
   public void testLoad() {
     Map<String, List<String>> result = ResourceLoader.loadFromClasspath();
@@ -25,10 +23,8 @@ public class ResourceLoaderTest {
     Map<String, List<String>> resultWithSpring = ResourceLoader.loadFromClasspath();
     assertEquals(result.size() + 1, resultWithSpring.size());
     assertNotNull(resultWithSpring.get(TEST_JAR));
-    // TODO review the generated test code and remove the default call to fail.
   }
 
-  /** Test of loadFromURL method, of class ClasspathResourceUtil. */
   @Test
   public void testLoadFromUrl_Jar() throws MalformedURLException {
     String testFile = TEST_JAR;
@@ -37,7 +33,6 @@ public class ResourceLoaderTest {
     assertEquals(expResult, result);
   }
 
-  /** Test of loadFromURL method, of class ClasspathResourceUtil. */
   @Test
   public void testLoadFromUrl_Zip() throws MalformedURLException {
     String testFile = TEST_ZIP;