+++ ChangeLog.md
@@ -1,5 +1,9 @@
-# WALib - common library Copyright (c) 2020-2021
+# WALib - common library Copyright (c) 2020-2021, 2023
 Author: Walery Wysotsky <dev@wysotsky.info>
+
+## [0.5.0] - 2023-03-31
+### Added
+- db package - Common database info
 
 ## [0.4.0] - 2021-03-08
 ### Added